Move age group calculation for changesets to its own module
